+++ packages/project-access/test/project/minUI5Version.test.ts
@@ -32,13 +32,41 @@ describe('getters for minUI5Version', () => {
         });
     });
 
-    describe('getMinUI5VersionAsArray', () => {
+    describe('getMinUI5VersionAsArray, with validation', () => {
+        const testVersions: Array<TestCase> = [
+            { minUI5Version: undefined, expectedResult: [] },
+            { minUI5Version: 'a.b.c', expectedResult: [] },
+            { minUI5Version: ['a.b.c'], expectedResult: [] },
+            { minUI5Version: '1.76.32', expectedResult: ['1.76.32'] },
+            { minUI5Version: '1.76', expectedResult: [] },
+            { minUI5Version: '1.76-snapshot', expectedResult: [] },
+            { minUI5Version: ['1.120.3'], expectedResult: ['1.120.3'] },
+            { minUI5Version: ['1.125'], expectedResult: [] },
+            { minUI5Version: ['1.120.13', 'a.b.c.'], expectedResult: ['1.120.13'] }
+        ];
+
+        testVersions.forEach((testCase) => {
+            test(`getMinUI5VersionAsArray: minUI5Version = ${testCase.minUI5Version}`, () => {
+                const manifest = {
+                    'sap.ui5': {
+                        dependencies: {
+                            minUI5Version: testCase.minUI5Version
+                        }
+                    }
+                } as Manifest;
+                expect(getMinUI5VersionAsArray(manifest)).toEqual(testCase.expectedResult);
+            });
+        });
+    });
+
+    describe('getMinUI5VersionAsArray, no validation', () => {
         const testVersions: Array<TestCase> = [
             { minUI5Version: undefined, expectedResult: [] },
             { minUI5Version: 'a.b.c', expectedResult: ['a.b.c'] },
             { minUI5Version: ['a.b.c'], expectedResult: ['a.b.c'] },
             { minUI5Version: '1.76.32', expectedResult: ['1.76.32'] },
             { minUI5Version: '1.76', expectedResult: ['1.76'] },
+            { minUI5Version: '1.76-snapshot', expectedResult: ['1.76-snapshot'] },
             { minUI5Version: ['1.120.3'], expectedResult: ['1.120.3'] },
             { minUI5Version: ['1.125'], expectedResult: ['1.125'] },
             { minUI5Version: ['1.120.13', 'a.b.c.'], expectedResult: ['1.120.13', 'a.b.c.'] }
@@ -53,7 +81,7 @@ describe('getters for minUI5Version', () => {
                         }
                     }
                 } as Manifest;
-                expect(getMinUI5VersionAsArray(manifest)).toEqual(testCase.expectedResult);
+                expect(getMinUI5VersionAsArray(manifest, true)).toEqual(testCase.expectedResult);
             });
         });
     });